# server/openapi.yaml
openapi: 3.0.3
info:
  title: TinyIce API
  description: |
    REST API for TinyIce audio streaming server.

    ## Authentication
    Two authentication methods are supported:

    **Session cookie** — Log in via `POST /login` to get a `sid` session cookie.

    **Bearer token** — Create an API token in the admin panel (`/admin/tokens`) and pass it
    as `Authorization: Bearer ti_...` header. Tokens can have optional expiry dates and
    track last-used time/IP.

    No CSRF tokens are needed for JSON API requests — the `Content-Type: application/json`
    header provides implicit CSRF protection since it triggers a CORS preflight for
    cross-origin requests.

    ## Real-time Updates
    Use the SSE endpoint `GET /events` for real-time metadata, stream status, and
    listener count updates without polling.
  version: "2.0"
  license:
    name: MIT
servers:
  - url: /
    description: This server
...
paths:
...
  /api/branding/logo:
    post:
      tags: [Branding]
      summary: Upload logo image
      description: Upload a PNG, JPG, or SVG logo (max 2MB). Saved to `branding/logo.{ext}`.
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [logo]
              properties:
                logo:
                  type: string
                  format: binary
                  description: Image file (PNG, JPG, SVG)
      responses:
        "200":
          description: Logo uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  path:
                    type: string
  /branding/logo:
    get:
      tags: [Branding]
      summary: Get logo image
      description: Serves the uploaded logo file. Returns 404 if no logo is set.
      responses:
        "200":
          description: Logo image
          content:
            image/*: {}
        "404":
          description: No logo configured

  # ── Settings ──────────────────────────────────────────────
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: sid
      description: Session cookie obtained via POST /login
    bearer:
      type: http
      scheme: bearer
      description: API token created in the admin panel (format ti_...)
...